src/requirements.txt
gunicorn==19.4.5
celery==3.1.19
redis==2.10.5
psycopg2==2.6.1
pillow==3.1.0
querystring-parser==1.2.3
django==1.9.1
djangorestframework==3.3.2
django-import-export==0.4.2
django-reversion==1.10
django-guardian==1.4.1
django-mptt==0.8.0
django-ckeditor==5.0.2
django_select2==5.5.0
django-bootstrap3==6.2.2
django-axes==1.6.0
#django-user-sessions==1.3.0
